# Use BenchmarkDotNet for performance testing, add protobuf-net for compare