--- a/Bond.PerformanceTest/BenchmarksConfig.cs
+++ b/Bond.PerformanceTest/BenchmarksConfig.cs
@@ -1,0 +1,18 @@
+﻿using BenchmarkDotNet.Columns;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Exporters;
+using BenchmarkDotNet.Exporters.Csv;
+
+namespace Bond.PerformanceTest
+{
+    public class BenchmarksConfig : ManualConfig
+    {
+        public BenchmarksConfig()
+        {
+            Add(RankColumn.Arabic);
+            Add(StatisticColumn.OperationsPerSecond);
+            Add(CsvMeasurementsExporter.Default);
+            Add(RPlotExporter.Default);
+        }
+    }
+}