--- a/Bond.PerformanceTest/Settings.cs
+++ b/Bond.PerformanceTest/Settings.cs
@@ -1,9 +1,0 @@
-﻿namespace Bond.PerformanceTest
-{
-    public static class Settings
-    {
-        public const int Times = 1000000;
-
-        public const int WarmUp = 100;
-    }
-}