--- a/Bond.PerformanceTest/packages.config
+++ b/Bond.PerformanceTest/packages.config
@@ -1,7 +1,26 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="BenchmarkDotNet" version="0.10.2" targetFramework="net451" />
+  <package id="BenchmarkDotNet.Core" version="0.10.2" targetFramework="net451" />
+  <package id="BenchmarkDotNet.Toolchains.Roslyn" version="0.10.2" targetFramework="net451" />
   <package id="Bond.CSharp" version="3.0.4" targetFramework="net451" />
   <package id="Bond.Runtime.CSharp" version="3.0.4" targetFramework="net451" />
   <package id="Google.ProtocolBuffers" version="2.4.1.521" targetFramework="net451" />
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net451" />
+  <package id="Microsoft.CodeAnalysis.Common" version="1.3.2" targetFramework="net451" />
+  <package id="Microsoft.CodeAnalysis.CSharp" version="1.3.2" targetFramework="net451" />
   <package id="Newtonsoft.Json" version="6.0.5" targetFramework="net451" />
+  <package id="protobuf-net" version="2.1.0" targetFramework="net451" />
+  <package id="System.Collections" version="4.0.0" targetFramework="net451" />
+  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net451" />
+  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net451" />
+  <package id="System.Globalization" version="4.0.0" targetFramework="net451" />
+  <package id="System.Linq" version="4.0.0" targetFramework="net451" />
+  <package id="System.Reflection.Metadata" version="1.2.0" targetFramework="net451" />
+  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net451" />
+  <package id="System.Runtime" version="4.0.0" targetFramework="net451" />
+  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net451" />
+  <package id="System.Threading" version="4.0.0" targetFramework="net451" />
+  <package id="System.Threading.Tasks" version="4.0.0" targetFramework="net451" />
+  <package id="System.Threading.Tasks.Extensions" version="4.0.0" targetFramework="net451" />
 </packages>