… Bond.PerformanceTest/BondCompact.cs
… Bond.PerformanceTest/BondFast.cs
… Bond.PerformanceTest/MeasurementEx.cs
… Bond.PerformanceTest/Protobuf.cs
… person_proto_net.cs