Hi!
Thanks for the benchmark!

In this PR I suggest to use best benchmarking library - [BenchmarkDotNet](http://benchmarkdotnet.org/).
It [doing benchmarking right way](http://benchmarkdotnet.org/HowItWorks.htm). with many features like different exporters and reports.

Also, as I've use [protobuf-net from Marc Gravell](https://github.com/mgravell/protobuf-net) library in my daily work, I've added it for comparison.
And the results are surprised me :) 

For serialization:

Method |        Mean |     StdDev |       Op/s | Rank |  Gen 0 | Allocated |
--------------- |------------ |----------- |----------- |----- |------- |---------- |
GoogleProtobuf | 597.9805 ns | 14.5617 ns | 1672295.35 |    1 | 0.0524 |     324 B |
ProtobufNet | 874.6373 ns | 36.0717 ns | 1143331.11 |    4 | 0.0946 |     584 B |
CompactBond | 648.5406 ns | 22.0672 ns | 1541923.61 |    2 | 0.0453 |     288 B |
FastBond | 671.7614 ns |  7.5169 ns | 1488623.83 |    3 | 0.0466 |     288 B |

and for deserialization:

Method |          Mean |     StdDev |       Op/s | Rank |  Gen 0 | Allocated |
--------------- |-------------- |----------- |----------- |----- |------- |---------- |
GoogleProtobuf |   905.8166 ns | 30.9835 ns | 1103976.18 |    3 | 0.1165 |     684 B |
ProtobufNet | 1,370.8720 ns | 35.8331 ns |   729462.7 |    4 | 0.0704 |     508 B |
CompactBond |   559.1139 ns | 10.9851 ns | 1788544.25 |    1 | 0.0848 |     464 B |
FastBond |   569.7601 ns | 11.8539 ns | 1755124.52 |    2 | 0.0858 |     464 B |

The configuration for tests was:

``` ini
BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4790 CPU 3.60GHz, ProcessorCount=8
Frequency=3507508 Hz, Resolution=285.1027 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1586.0
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.6.1586.0
```

and here is performance charts, generated by BenchmarkDotNet for serialization:

![serializationbenchmarks-barplot](https://cloud.githubusercontent.com/assets/1401048/22469082/8de43d84-e7ec-11e6-9b6e-b50eb09fe7f4.png)

and for deserialization:

![deserializationbenchmarks-barplot](https://cloud.githubusercontent.com/assets/1401048/22469087/9233b680-e7ec-11e6-941c-b33d6e422eb1.png)

It seems like we can always use Compact Bond serializer - it outperforms Fast Bond Serializer and slightly worse in serialization scenario. Unfortunately, protobuf-net is lastest in the list (thanks for Google developers).

Hope, you consider to accept PR.