@takemyoxygen 2017-03-01
Thanks for the PR, I had a quick look and it indeed looks useful. I hope I'll find some time in the coming weeks to look in more details and merge the PR